refactor: centralize scrollbar UI and fix navigation
