+++ application-system/src/main/java/com/group_9/project/AccountAddressPage.java
@@ -23,6 +23,7 @@ public class AccountAddressPage extends Template {
 
     public AccountAddressPage() {
         BaseFrameSetup.applyAppIcon(this);
+        setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
         BackgroundPanel background = BaseFrameSetup.setupCompleteFrame(this, 3);
         
         JPanel sidebar = AccountSidebarUtil.createSidebar(this, "My Address");
